% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed.R
\name{func}
\alias{func}
\title{triple mean}
\usage{
func(X, Y, Z)
}
\arguments{
\item{X}{vector length n}

\item{Y}{vector length n}

\item{Z}{vector length n}
}
\value{
double
}
\description{
triple mean
}
\examples{
 func(seq(1,10,length = 50), seq(1,20,length = 50),seq(1,30,length = 50))
}
