% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified.R
\name{modified_estimator}
\alias{modified_estimator}
\title{Modified estimator}
\usage{
modified_estimator(X, Y, coef_vec = NULL)
}
\arguments{
\item{X}{matrix}

\item{Y}{vector}

\item{coef_vec}{a vector}
}
\value{
double
}
\description{
Modified estimator
}
\examples{
X <- matrix(rnorm(15),3,5)
Y <- c(2, 4.2, 5.8)
coef_vec <- c(1,2,3)
modified_estimator(X,Y,coef_vec)
}
