% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naive.R
\name{naive}
\alias{naive}
\title{calculate the Naive estimator}
\usage{
naive(X, Y)
}
\arguments{
\item{X}{numeric matrix}

\item{Y}{numeric vector}
}
\value{
double
}
\description{
calculate the Naive estimator
}
\examples{
X<-matrix(rnorm(15),3,5)
Y<-c(2, 4.2, 5.8)
naive(X,Y)
}
