% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed.R
\name{proposed_estimator}
\alias{proposed_estimator}
\title{proposed estimator}
\usage{
proposed_estimator(X, Y, coef_vec)
}
\arguments{
\item{X}{matrix}

\item{Y}{vector}

\item{coef_vecvector}{a vector}
}
\value{

}
\description{
proposed estimator
}
\examples{
X <- matrix(rnorm(15),3,5)
Y <- c(2, 4.2, 5.8)
coef_vec <- c(1,2,3)
proposed_estimator(X,Y,coef_vec)
}
