% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposed.R
\name{selection_algoritm}
\alias{selection_algoritm}
\title{Selection algorithm}
\usage{
selection_algoritm(X, Y)
}
\arguments{
\item{X}{matrix}

\item{Y}{vector}
}
\value{

}
\description{
Selection algorithm
}
\examples{
X<-matrix(rnorm(15),3,5)
Y<-c(2, 4.2, 5.8)
selection_algoritm(X,Y)
}
